Ids bug fixed

+++ src/main/java/utils/HibernateUtil.java
@@ -147,7 +147,7 @@ public class HibernateUtil {
     /**
      * Update the existing User.
      */
-    public static void updateUser(int id, String login) {
+    public static void updateUser(long id, String login) {
         // Create an EntityManager and EntityTransaction
         EntityManager manager = ENTITY_MANAGER_FACTORY.createEntityManager();
         EntityTransaction transaction = null;
